requirements.txt
ads1015==0.0.8
aniso8601==9.0.1
astral==3.2
astroid==2.15.6
blinker==1.6.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
dill==0.3.7
enviroplus==0.0.6
Flask==2.3.2
Flask-RESTful==0.3.10
font-roboto==0.0.1
fonts==0.0.3
i2cdevice==0.0.7
idna==3.4
importlib-metadata==6.8.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.9.0
ltr559==0.1.1
MarkupSafe==2.1.3
mccabe==0.7.0
numpy==1.25.2
paho-mqtt==1.6.1
Pillow==10.0.0
pimoroni-bme280==0.1.1
platformdirs==3.10.0
pms5003==0.0.5
pycparser==2.21
pylint==2.17.5
pyserial==3.5
pytz==2023.3
requests==2.31.0
RPi.GPIO==0.7.1
six==1.16.0
smbus==1.1.post2
sounddevice==0.4.6
spidev==3.6
ST7735==0.0.4.post1
tomli==2.0.1
tomlkit==0.12.1
typing-extensions==4.7.1
urllib3==2.0.4
Werkzeug==2.3.6
wrapt==1.15.0
zipp==3.16.2
